40% less code.
Refactored the inner workings and simplified how things are stored. There was no need for the complex checking of the arguments for #record, because in the end they were outputted the same. Same goes for #identity.

The HTML sanitation could go too, since #to_json takes care of all that.